pyproject.toml
[project]
name = "tg-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "celery-redbeat>=2.3.2",
    "dishka>=1.6.0",
    "g4f>=0.5.3.2",
    "google-auth>=2.40.3",
    "gspread-asyncio>=2.0.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "oauth2client>=4.1.3",
    "pydantic-settings>=2.9.1",
    "pydub>=0.25.1",
    "pytz>=2025.2",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "telethon>=1.40.0",
]
